restructuring so dbConnect isn't in API
