+++ pages/api/auth/[...nextauth].ts
@@ -1,6 +1,6 @@
 import NextAuth from 'next-auth'
 import CredentialProvider from 'next-auth/providers/credentials'
-import { dbConnect } from '../_dbConnect'
+import { dbConnect } from '../../../server/_dbConnect'
 
 /* eslint-disable */
 export default NextAuth({
